further rearranging of deck chairs in MU::Cloud so we can get a better CodeClimate grade later
