+++ modules/mu/cloud/server_pool.rb
@@ -1,0 +1,1 @@
+server.rb
